requirements (2).txt
torch==2.4.1
transformers==4.33.2
faiss-cpu==1.9.0.post1  
numpy==1.26.4
gradio==5.9.1
matplotlib==3.7.1
jsonschema==4.23.0
